+++ MarioCollision.cpp
@@ -344,9 +344,9 @@ bool Mario::OnCollisionWithFlagPole(const CollisionEvent* collisionEvent)
 		return false;
 
 
-	float score = 0;
-	float bottom = flagPole->GetBoundingBox().bottom;
-	float touchingPoint = position.y + (power == MarioPower::Normal ? 16 : 32); // Mario's feet position
+	int score = 0;
+	float bottom = static_cast<float>(flagPole->GetBoundingBox().bottom);
+	float touchingPoint = position.y + (power == MarioPower::Normal ? 16.0f : 32.0f); // Mario's feet position
 	float touchingHeight = bottom - touchingPoint;
 
 	score = GetFlagBonusScore(touchingHeight);
@@ -359,7 +359,7 @@ bool Mario::OnCollisionWithFlagPole(const CollisionEvent* collisionEvent)
 	AudioManager::GetInstance()->PlaySFX(FLAG_PULL);
 
 	// set state to flag sliding
-	velocity = Vector2(0, 0);
+	velocity = Vector2(0.0f, 0.0f);
 	state = MarioState::PullingFlag;
 	// make mario face right
 	isFacingRight = true;
@@ -369,11 +369,11 @@ bool Mario::OnCollisionWithFlagPole(const CollisionEvent* collisionEvent)
 	// initiate sliding down
 	auto snapPosition = flagPole->GetSnapPosition();
 	position.x = snapPosition.x;
-	position.y = max(position.y, snapPosition.y);
+	position.y = max(position.y, static_cast<float>(snapPosition.y));
 
-	float offset = power == MarioPower::Normal ? 16 : 32;
+	float offset = power == MarioPower::Normal ? 16.0f : 32.0f;
 
-	slidingToYWinning = flagPole->GetBoundingBox().bottom - offset;
+	slidingToYWinning = static_cast<float>(flagPole->GetBoundingBox().bottom) - offset;
 	marioWinningMoveToPosition = flagPole->GetMoveToPosition();
 	return true;
 }
